requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
certifi==2023.7.22
charset-normalizer==3.2.0
dj-database-url==2.1.0
Django==4.2.7
django-admin-interface==0.23.0
django-ckeditor==6.7.0
django-colorfield==0.10.1
django-countries==7.5.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-health-check==3.17.0
django-js-asset==2.1.0
django-quill-editor==0.1.40
django-watchman==1.3.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
gunicorn==21.2.0
idna==3.4
openpyxl==3.1.2
packaging==23.1
Pillow==10.0.1
psutil==5.9.6
PyMySQL==1.1.0
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
sentry-sdk==1.37.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.7.1
urllib3==2.0.4
whitenoise==6.5.0
